Add Recents Emoji

+++ emoji-keyboard/src/main/java/inc/flide/android/emoji_keyboard/sqlite/EmojiSQLiteHelper.java
@@ -8,17 +8,19 @@ public class EmojiSQLiteHelper extends SQLiteOpenHelper {
 
     public static final String TABLE_RECENTS = "recent";
     public static final String COLUMN_ID = "_id";
-    public static final String COLUMN_TEXT = "text";
-    public static final String COLUMN_ICON = "icon";
-    public static final String COLUMN_COUNT = "count";
+    public static final String COLUMN_UNICODE_HEX_CODE = "Unicode_Hex_Code";
+    public static final String COLUMN_CATEGORY = "Category";
+    public static final String COLUMN_COUNT = "Count";
 
-    private static final String DATABASE_NAME = "recent.db";
-    private static final int DATABASE_VERSION = 1;
+    private static final String DATABASE_NAME = "RecentEmojis.db";
+    private static final int DATABASE_VERSION = 2;
 
     // Database creation sql statement
     private static final String DATABASE_CREATE = "CREATE TABLE " + TABLE_RECENTS + "("
-            + COLUMN_ID + " INTEGER PRIMARY KEY," + COLUMN_TEXT + " TEXT,"
-            + COLUMN_ICON + " TEXT," + COLUMN_COUNT + " INTEGER" + ")";
+                                                    + COLUMN_ID + " INTEGER PRIMARY KEY,"
+                                                    + COLUMN_UNICODE_HEX_CODE + " TEXT,"
+                                                    + COLUMN_CATEGORY + " TEXT,"
+                                                    + COLUMN_COUNT + " INTEGER" + ")";
 
     public EmojiSQLiteHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -31,8 +33,10 @@ public class EmojiSQLiteHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_RECENTS);
-        onCreate(db);
+        if (newVersion > oldVersion) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_RECENTS);
+            onCreate(db);
+        }
     }
 
 }
